refactor(repo): mv src to pyprql
Renames src to pyprql and makes neceessary changes in pyproject.toml and tests to reflect this. This brings in-line with standard packaging practices, and prevents the need for PYTHONPATH variables.
