+++ pyprql/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+"""Source code for PyPrql."""
+__version__ = "0.0.0"
